The MSVC linker doesn't have a -L option.
The MSVC linker doesn't have a -L option, and doesn't understand
posix paths.  Move paths to an envvar with a prefix (-LIBPATH:).

* libltdl/m4/libtool.m4: Add new tag variables to specify
to what variable (dashL_envvar) -L options should be moved,
and what to prepend to the -L path (dashL_envvar_spec).
* libltdl/config/ltmain.m4sh (func_dashL_to_envvar): New
function that moves -L options to a variable.
* libltdl/config/ltmain.m4sh (func_mode_link): Prior to
linking, call above function if dashL_envvar is set.