fixup: restore EXPORTS test
Commit v2.4.2-120-g962aa91
syntax-check: fix violations and implement sc_prohibit_test_const_follows_var
inadvertedly reversed the meaning of the comparison.

* build-aux/ltmain.m4sh (func_mode_link) [cygwin|mingw|cegcc]: Restore
the EXPORTS test.  We need to look at the symbols when it's _not_
already a .def file (in which case we trust the user input blindly).

Copyright-paperwork-exempt: Yes
Signed-off-by: Peter Rosin <peda@lysator.liu.se>