testsuite: with SunStudio12.x f77, don't fail on system symbols
Modern SunStudio f77 is just a wrapper for f90, which
explains why current master fails these test groups with:

 Undefined	first referenced
  symbol  	    in file
 __f90_slw_ch      .libs/foof2.o
 __f90_sslw        .libs/foof2.o
 __f90_eslw        .libs/foof2.o
 ld: fatal: symbol referencing errors. No output written to
.libs/libfoo2.so.0.0.0

So we allow some undefined symbols in the libraries, which are
resolved by the $(FLIBS) added to the final binaries in the tests.
This seems like the simpler, and less error inducing solution to
me (compared to adding $(FLIBS) to every library LDFLAGS line).
* tests/f77demo.at (libfoo_la_LDFLAGS, libfoo2_la_LDFLAGS)
(libfoo3_la_LDFLAGS, libmix_la_LDFLAGS): Remove -no-undefined
flag.
* tests/fcdemo.at: Likewise.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>