ltmain.in: ensure old libraries are not archived
Previously, adding a static archive library, aka. old library, as a
dependency of a new libtool library compiled as a static archive (be it
a convenience library or an installed static library) resulted in the
dependency archive being added into the target archive as-is.

However, linkers are usually not able to use nested archive as-is. It
results in errors like these on Linux and Darwin at least:

    /usr/bin/ld: ./.libs/libfoo.a: member ./.libs/libfoo.a(libbar.a) in
    archive is not an object

or

    /usr/bin/nm: libbar.a: file format not recognized

With libfoo.a being the static libtool target and libbar.a being the
nested/linked static library.

Libtool has a quite close behaviour with convenience libraries, where
the content (objects) of the convenience library are extracted and added
back into the new target, but that´s mostly a libtool creation.

The static dependencies should not be "linked" into other static
libraries as there should not be linking involved, and we don´t want to
potentially duplicate symbols like done for the convenience libraries.

Instead, keep them in the dependency_libs of the new libtool archive
created and forward them until we link a shared library or an
executable. We do need to ensure they are referenced with absolute path
since we can link from a different folder than the one it was referenced
from.

We also need to filter out the shared library case since we don't want
to include the static archives in dependencies again.

The test was marked XFAIL since 2010 because the fix for the
nested-archive bug was incomplete. Now that ltmain.in properly
forwards old library dependencies instead of archiving them,
the test passes: remove the AT_XFAIL_IF and correct the
foo.$OBJEXT assertion (objects from the dependency must NOT
appear in the target archive).

Absolute paths are forwarded as-is like above.
Relative paths are transformed into absolute using pwd.

The old deplibs are not appended if they have been linked
into a shared library output already.

Reported: https://savannah.gnu.org/patch/?10569