libtool: port better to POSIX grep
This fixes some other regular expressions that GNU grep 3.8 does not
warn about (I found them with an even-pickier 'grep').

I did not look systematically for all portability bugs of this form.

Url: https://savannah.gnu.org/patch/index.php?10282

* tests/cdemo.at (allow_undefined_flag):
Prefer } to \} in EREs, as \} is not portable.
* tests/libtool.at (quote shell meta-characters in filenames):
Prefer ] to \] in BREs, as \] is not portable.